【资源管理重构】代码整理
