+++ assets/Script/res/ResLoader.ts.meta
@@ -1,6 +1,6 @@
 {
   "ver": "1.0.8",
-  "uuid": "6bb26b58-3dd7-464b-b61f-7921b26ac2fc",
+  "uuid": "4b0811b5-6fda-49bd-b0ed-56df405215dd",
   "isPlugin": false,
   "loadPluginInWeb": true,
   "loadPluginInNative": true,
