--- assets/Script/res/ResLoader2.ts.meta
@@ -1,9 +1,0 @@
-{
-  "ver": "1.0.8",
-  "uuid": "4b0811b5-6fda-49bd-b0ed-56df405215dd",
-  "isPlugin": false,
-  "loadPluginInWeb": true,
-  "loadPluginInNative": true,
-  "loadPluginInEditor": false,
-  "subMetas": {}
-}
